requirements.txt
selenium
psycopg2
supabase
asyncio
apify_client
google-image-source-search
fuzzywuzzy
pandas
python-Levenshtein
chromedriver-autoinstaller
pyvirtualdisplay
webdriver-manager
parsel
requests
scikit-image
numpy
opencv-python
matplotlib
Pyarrow
imapclient
tempfile
imaplib
email
re
glob
unicodedata
decimal
